+++ src/Implicit/SharedObjectPools.cs
@@ -4,6 +4,6 @@ namespace Implicit
 {
     internal static class SharedObjectPools
     {
-        public static ObjectPoolSlim<List<KeyValuePair<string, double>>> KeyValueLists { get; } = new(() => new(), x => x.Clear());
+        public static ObjectPoolSlim<List<KeyValuePair<string, float>>> KeyValueLists { get; } = new(() => new(), x => x.Clear());
     }
 }
